Fix: microservices Bad URL Error
When running in microservices mode the client recieves the error
400: Bad URL localhost:3000/employees

This looks to be caused by posts to addNewEmployee being made without
a last name for the employee.

Add logic similar to loader.js from monolithic code to db_loader_service/controllers/loader.js
to ensure there enough last names for all the users to be added.